requirements.txt
pycryptodome==3.15.0
pywin32==302
